# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Currency API",
    "version": "1.0.0",
    "description": "API for managing currencies"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/currency": {
      "post": {
        "summary": "Create currency",
        "tags": ["Currency"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCurrency"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Currency created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Currency"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/currency/{id}": {
      "get": {
        "summary": "Get currency by id",
        "tags": ["Currency"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Currency found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Currency"
                }
              }
            }
          },
          "404": {
            "description": "Currency not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update currency",
        "tags": ["Currency"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCurrency"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Currency updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Currency"
                }
              }
            }
          },
          "404": {
            "description": "Currency not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete currency",
        "tags": ["Currency"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Currency deleted"
          },
          "404": {
            "description": "Currency not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/history/{currency}": {
      "get": {
        "tags": ["History"],
        "summary": "Get history",
        "parameters": [
          {
            "name": "currency",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "EUR"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Currency found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceHistoryList"
                }
              }
            }
          },
          "404": {
            "description": "Currency not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/price/{currency}": {
      "get": {
        "tags": ["Price"],
        "summary": "Get price",
        "parameters": [
          {
            "name": "currency",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "EUR"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Currency found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceItem"
                }
              }
            }
          },
          "404": {
            "description": "Currency not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PriceHistoryItem": {
        "type": "object",
        "required": ["symbol"],
        "properties": {
          "symbol": {
            "type": "object",
            "required": ["price", "date"],
            "properties": {
              "date": {
                "type": "string",
                "description": "TimeStamp",
                "example": "123343443"
              },
              "price": {
                "type": "string",
                "description": "Price",
                "example": "0.01675"
              }
            }
          }
        }
      },
      "PriceItem": {
        "type": "object",
        "required": ["symbol", "price"],
        "properties": {
          "symbol": {
            "type": "string",
            "description": "Ticker pair",
            "example": "EURETH"
          },
          "price": {
            "type": "string",
            "description": "Price",
            "example": "0.01675"
          }
        }
      },
      "PriceList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PriceItem"
        }
      },
      "PriceHistoryList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PriceHistoryItem"
        }
      },
      "Currency": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Bitcoin"
          },
          "ticker": {
            "type": "string",
            "example": "BTC"
          }
        },
        "required": ["id", "name", "ticker"]
      },
      "CreateCurrency": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Bitcoin"
          },
          "ticker": {
            "type": "string",
            "example": "BTC"
          }
        },
        "required": ["name", "ticker"]
      },
      "UpdateCurrency": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ethereum"
          },
          "ticker": {
            "type": "string",
            "example": "ETH"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Currency not found"
          }
        },
        "required": ["error"]
      }
    }
  }
}